requirements.txt
# XLearning-Agent Requirements
# Python >= 3.10

# ============================================
# LangChain 核心（简历关键词 ✔️）
# ============================================
langchain==0.3.17
langchain-core==0.3.31
langchain-community==0.3.16
langchain-text-splitters==0.3.5
langchain-chroma==0.2.2

# ============================================
# LangGraph（Checkpoint 2 后必做）
# ============================================
langgraph==0.2.68

# ============================================
# LangSmith 可观测性（必做）
# ============================================
langsmith==0.2.11

# ============================================
# 向量存储
# ============================================
chromadb==0.6.3

# ============================================
# Tongyi/DashScope（主要 Provider）
# ============================================
dashscope==1.22.1

# ============================================
# UI
# ============================================
streamlit==1.41.1

# ============================================
# PDF 解析
# ============================================
pymupdf==1.25.3

# ============================================
# 工具库
# ============================================
python-dotenv==1.0.1
pydantic==2.10.6
rich==13.9.4
httpx==0.28.1
